backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
python-multipart
python-dotenv
langchain
langchain-huggingface
langchain-core
huggingface_hub